requirements.txt
pydantic-ai==1.4.0
python-dotenv==1.0.0
httpx==0.28.1
mcp-tavily
pyyaml>=6.0